:hammer: [layout] 抽象部分通用方法到单独文件，并调整部分类型
