+++ src/shared/layout/types.ts
@@ -1,0 +1,1 @@
+export type StateDataType = number | string | null;
